adapt unit test
